Refactor tests;
